if image_width and image_height are given fields, then set image_tile_size accordingly and scale
The full image can be seen as tiles if image_width and image_height are
set in the properties file. In that case, image_tile_size is
overwritten and set to the minimum of image_width and image_height.
Also, the scale is set to 1.0/image_tile_size.

If no image_width and image_height, image_tile_size is set to a default
of 500 pixels.

Setting property fields in  dbconnect.py doesn’t do anything; property
fields must be set in properties.py.