networkx v1 -> v3 changes (part 2)
Includes the .weakly_connected_component_subgraphs function being
changed to weakly_connected_components which returns a reference to a
subgraph rather than the subgraph itself.

Also handles the changes to .in_degree and .out_degree which now return
a generator rather than a dict. This is compounded by changes to the
.keys() and .values() attributes of a dict, which no longer return
lists.